Struct pop3::POP3Stream

pub struct POP3Stream {
    pub is_authenticated: bool,
    // some fields omitted
}

The stream to use for interfacing with the POP3 Server.

Fields
is_authenticated: bool
Methods

impl POP3Stream

fn connect<A: ToSocketAddrs>(
addr: A,
ssl_context: Option<SslConnector>,
domain: &str
) -> Result<POP3Stream>

Creates a new POP3Stream.

fn login(&mut self, username: &str, password: &str) -> POP3Result

Login to the POP3 server.

fn stat(&mut self) -> POP3Result

Gives the current number of messages in the mailbox and the total size in bytes of the mailbox.

fn uidl(&mut self, message_number: Option<i32>) -> POP3Result

fn list(&mut self, message_number: Option<i32>) -> POP3Result

List displays a summary of messages where each message number is shown and the size of the message in bytes.

fn retr(&mut self, message_id: i32) -> POP3Result

retrieves the message of the message id given.

fn dele(&mut self, message_id: i32) -> POP3Result

Delete the message with the given message id.

fn rset(&mut self) -> POP3Result

This resets the session to its original state.

fn quit(&mut self) -> POP3Result

Quits the current session.

fn noop(&mut self) -> POP3Result

Doesn't do anything. This is usually just used to keep the connection open.